requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
distlib==0.3.6
Django==4.2.1
django-cleanup==9.0.0
django-cors-headers==4.3.1
django-multiupload==0.6.1
django-paypal==2.0
djangorestframework==3.14.0
filelock==3.10.0
idna==3.4
pbr==5.11.1
Pillow==9.5.0
platformdirs==3.1.1
pycodestyle==2.10.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
stevedore==5.0.0
stripe==5.4.0
urllib3==2.0.2
virtualenv==20.21.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
